test_cw now searches through -4Mhz to +4Mhz, is compatible with Swift-NAP-HDL commit 7cfd2d770f9db1fb92b15c16c5aeaf04835c7ae8

+++ src/cw.c
@@ -23,7 +23,7 @@
 
 cw_state_t cw_state;
 
-void do_one_cw(s16 carrier_freq, corr_t* corrs)
+void do_one_cw(s32 carrier_freq, corr_t* corrs)
 {
   cw_write_init_blocking(carrier_freq);
